requirements.txt
black
pyyaml
tableauserverclient
tableauhyperapi
wheel